requirements.txt
django
django-silk
django-cors-headers
Pillow
django-debug-toolbar
drf-yasg
django-sslserver